+++ src/LayerTransition.cpp
@@ -1,7 +1,7 @@
 #include "Layer.hpp"
 #include <regex>
 
-extern std::map<short, ChromaSetup> setups;
+extern std::map<short, ChromaPattern> setups;
 extern std::map<std::string, bool> opts;
 extern float speed;