+++ src/include/mynode.cpp
@@ -1,6 +1,6 @@
 #include "mynode.hpp"
 
-extern std::map<short, ChromaSetup> setups;
+extern std::map<short, ChromaPattern> setups;
 extern std::map<std::string, bool> opts;
 
 // UnlockType tags by RobTop